Extract shared NewToPrairieLearnCard component
Deduplicate the "New to PrairieLearn?" card with documentation links
that was duplicated between the question creation page and the AI
question generation editor.
